Provided Create Class shortcut
Create class creates the basic java class structure based off of the filename and path.

Packaging comes from the path and is keyed off of the "java" folder as the cut-off point for the path.
For instance, the following path will produce the following path:
/Users/jsmith/project/myjavaproject/src/main/java/com/myproject/mymodels/MyModel.java
  package com.myproject.mymodels;

The class is keyed off of the filename.  The aforementioned example produces the following class:
public class MyModel {
  public MyModel() {
  }
}